fix worker

+++ src/InitializeData/useInitializeData.ts
@@ -1,24 +1,27 @@
 import { useEffect, useState } from 'react';
 import { useDispatch } from 'react-redux';
-import { setUsers } from '../store/usersSlice';
+import { setUsers } from '../store/usersSlice'; 
 
-export const useInitializeData = (count: number = 1000000) => {
+export const useInitializeData = () => {
   const dispatch = useDispatch();
   const [loading, setLoading] = useState(true);
+  const USER_COUNT = 7500000; 
 
   useEffect(() => {
     const worker = new Worker(new URL('../utils/dataGenerator.worker.ts', import.meta.url));
     
-    worker.postMessage({ count });
+    worker.postMessage({ count: USER_COUNT });
     
     worker.onmessage = (e) => {
-      dispatch(setUsers(e.data));
-      setLoading(false);
-      worker.terminate();
+      if (e.data) {
+        dispatch(setUsers(e.data)); 
+        setLoading(false);
+        worker.terminate();
+      }
     };
 
     return () => worker.terminate();
-  }, [dispatch, count]);
+  }, [dispatch]);
 
   return { loading };
 };
